Working PoC
The tool is officially "working".  There are a number of changes to be
made from here, but it's effectively working.

FYI, the serial port (and mechanism) should be taken into account.  If
you're not using a receipt printer with an integrated USB -> Serial
converter, you may need to change the initialization of the printer.
For reference, check out [python-escpos][1].

[1]: https://python-escpos.readthedocs.io/en/latest/